requirements.txt
librosa==0.11.0
matlabengine==24.2
numpy==1.23.4
pandas==1.5.1
pydub==0.25.1
pyloudnorm==0.1.1
PyQt6==6.10.1
pyqt6_sip==13.10.2
pyqtgraph==0.13.7
qt_material==2.17
scipy==1.16.3
sounddevice==0.5.3
soundfile==0.13.1
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
